Bammann1990Modeling

The following equations are those employed in the Bammann (1990) paper that are implemented in the Bammann1990Modeling type, constructor, and that method of the kernel function, update associate with this type. As such, the internal equations use the same nomenclature for plastic strain rate, $\mathbf{D}^{p} \equiv \dot{\epsilon}^{(p)}$; the second-rank, deviatoric tensors for Cauchy stress, $\underset{\sim}{\sigma}'$ and kinematic hardening, $\underset{\sim}{\alpha}'$ which is an ISV; the scalar isotropic hardening, $\kappa$, and the constants associated with the dynamic and static recovery temperature equations. Of major importance is that, although not explicitly listed in the publication, the equations for $h$ and $H$ are included in this implementation (c. f. DYNA3D User Manual (1993)).

\[\begin{aligned} % plastic strain rate \mathbf{D}^{p} &= f(\theta)\sinh\left[ \frac{ |\underset{\sim}{\xi}| - \kappa - Y(\theta) }{ V(\theta) } \right]\frac{\underset{\sim}{\xi}'}{|\underset{\sim}{\xi}'|}\text{, let }\underset{\sim}{\xi}' = \underset{\sim}{\sigma}' - \underset{\sim}{\alpha}' \\ % kinematic hardening \dot{\underset{\sim}{\alpha}} &= h\mu(\theta)\mathbf{D}^{p} - [r_{d}(\theta)|\mathbf{D}^{p}| + r_{s}(\theta)]|\underset{\sim}{\alpha}|\underset{\sim}{\alpha} \\ % isotropic hardening \dot{\kappa} &= H\mu(\theta)\mathbf{D}^{p} - [R_{d}(\theta)|\mathbf{D}^{p}| + R_{s}(\theta)]\kappa^{2} \\ % flow rule F &= |\underset{\sim}{\sigma}' - \underset{\sim}{\alpha}'| - \kappa - \beta(|\mathbf{D}^{p}|, \theta) \\ % initial yield stress beta \beta(\mathbf{D}^{p}, \theta) &= Y(\theta) + V(\theta)\sinh^{-1}\left(\frac{|\mathbf{D}^{p}|}{f(\theta)}\right) \\ V(\theta) &= C_{ 1} \exp\left( -\frac{ C_{ 2} }{ \theta } \right) \\ Y(\theta) &= C_{ 3} \exp\left( \frac{ C_{ 4} }{ \theta } \right) \\ f(\theta) &= C_{ 5} \exp\left( -\frac{ C_{ 6} }{ \theta } \right) \\ r_{d}(\theta) &= C_{ 7} \exp\left( -\frac{ C_{ 8} }{ \theta } \right) \\ r_{s}(\theta) &= C_{ 9} \exp\left( -\frac{ C_{10} }{ \theta } \right) \\ R_{d}(\theta) &= C_{11} \exp\left( -\frac{ C_{12} }{ \theta } \right) \\ R_{s}(\theta) &= C_{13} \exp\left( -\frac{ C_{14} }{ \theta } \right) \\ h(\theta) &= C_{15} \exp\left( \frac{ C_{16} }{ \theta } \right) \\ H(\theta) &= C_{17} \exp\left( \frac{ C_{18} }{ \theta } \right) \\ \end{aligned}\]

Types

BammannChiesaJohnsonPlasticity.Bammann1990Modeling — Type

Structure for viscoplasticity model with loading conditions and material properties. Here, uses the effective strain rate based on applied strain rate and loading direction.

BammannChiesaJohnsonPlasticity.Bammann1990Modeling — Method

Bammann1990Modeling(Ω, μ)

Outer constructor for loading conditions and material properties which assumes a Poisson's ratio of 0.5. Here, μ is the shear modulus.

Functions

BammannChiesaJohnsonPlasticity.update — Method

Using the equations and constants from Bammann (1990), this kernel function maps the current material state and ISVs onto the next configuration. Note: though not explicitly listed in paper, temperature equations h = C₁₅ * exp(C₁₆ * θ) and H = C₁₇ * exp(C₁₈ * θ) are included (c. f. DYNA3D User Manual (1993)).

References

[5]

D. J. Bammann. Modeling Temperature and Strain Rate Dependent Large Deformations of Metals. Applied Mechanics Reviews 43, S312-S319 (1990).

[6]

R. G. Whirley and B. Engelmann. DYNA3D: A Nonlinear, Explicit, Three-Dimensional Finite Element Code for Solid and Structural Mechanics, User Manual. Revision 1 (Lawrence Livermore National Lab.(LLNL), Livermore, CA (United States), 1993).

Index