OptimizationBCJPlasticityExt

This extension is meant to provide additional functionality to the base package with the intent of performing optimization studies with the Optimization.jl and LossFunctions.jl packages to calibrate material model constants against experimental data.

Types

ContinuumMechanicsBase.MaterialHistory — Type

Structure for storing the behavior of a material as it evolves in time. Design to be used in time-dependent models such as viscoelasticity.

Functions

ContinuumMechanicsBase.parameter_bounds — Method

Set lower bounds to zero for physical admissibility.

ContinuumMechanicsBase.AlmansiStrainTensor! — Method

AlmansiStrainTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.AlmansiStrainTensor — Method

AlmansiStrainTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.CauchyStressTensor! — Method

CauchyStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.CauchyStressTensor — Method

CauchyStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.DeformationGradientTensor! — Method

DeformationGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.DeformationGradientTensor — Method

DeformationGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor! — Method

FirstPiolaKirchoffStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.FirstPiolaKirchoffStressTensor — Method

FirstPiolaKirchoffStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.GreenStrainTensor! — Method

GreenStrainTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.GreenStrainTensor — Method

GreenStrainTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseDeformationGradientTensor! — Method

InverseDeformationGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseDeformationGradientTensor — Method

InverseDeformationGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor! — Method

InverseLeftCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.InverseLeftCauchyGreenDeformationTensor — Method

InverseLeftCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.I₁ — Method

I₁(T::AbstractMatrix) -> Any

$I_1 = \text{tr}(T)$

ContinuumMechanicsBase.I₂ — Method

I₂(T::AbstractMatrix) -> Any

$I_2 = \frac{1}{2}\left(\text{tr}(T)^2 - \text{tr}(T^2)\right)$

ContinuumMechanicsBase.I₃ — Method

I₃(T::AbstractMatrix) -> Any

$I_3 = \det{T}$

ContinuumMechanicsBase.J — Method

J(T::AbstractMatrix) -> Any

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor! — Method

LeftCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.LeftCauchyGreenDeformationTensor — Method

LeftCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.MaterialOptimizationProblem — Method

MaterialOptimizationProblem(ψ, test, u₀, ps, adb, loss)

Creates an OptimizationProblem for use in Optimization.jl to find the optimal parameters.

Arguments:

ψ: material model to use

test or tests: A single or vector of ::AbstractMaterialTests to use when fitting the parameters

u₀: Initial guess for parameters

ps: Any additional parameters for calling predict()

adb: Select differentiation type from ADTypes.jl. The type is automatically applied to the type of AD applied to the OptimizationProblem also.

loss: Loss function from LossFunctions.jl

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor! — Method

RightCauchyGreenDeformationTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.RightCauchyGreenDeformationTensor — Method

RightCauchyGreenDeformationTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor! — Method

SecondPiolaKirchoffStressTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.SecondPiolaKirchoffStressTensor — Method

SecondPiolaKirchoffStressTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.StrainEnergyDensity! — Method

StrainEnergyDensity!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.StrainEnergyDensity — Method

StrainEnergyDensity(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.VelocityGradientTensor! — Method

VelocityGradientTensor!(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.VelocityGradientTensor — Method

VelocityGradientTensor(Model, State, Parameters; kwargs...)

ContinuumMechanicsBase.parameter_bounds — Method

parameter_bounds(
    _::ContinuumMechanicsBase.AbstractMaterialModel,
    _::ContinuumMechanicsBase.AbstractMaterialTest
) -> @NamedTuple{lb::@NamedTuple{C₁::Float64, C₂::Float64, C₃::Float64, C₄::Float64, C₅::Float64, C₆::Float64, C₇::Float64, C₈::Float64, C₉::Float64, C₁₀::Float64, C₁₁::Float64, C₁₂::Float64, C₁₃::Float64, C₁₄::Float64, C₁₅::Float64, C₁₆::Float64, C₁₇::Float64, C₁₈::Float64}, ub::Nothing}

Default bounds for each parameter in tuple for optimization. Without method dispatching for certain models, constants can be "optimized" in the range [-∞, ∞]. Use discretion whether this is realistic.

ContinuumMechanicsBase.parameters — Method

parameters(
    _::ContinuumMechanicsBase.AbstractMaterialModel
) -> NTuple{18, Symbol}

-ψ: Material Model

Returns:

Tuple of symbols for material model parameters

ContinuumMechanicsBase.predict — Method

predict(
    ψ::ContinuumMechanicsBase.AbstractMaterialModel,
    test::ContinuumMechanicsBase.AbstractMaterialTest,
    ps;
    kwargs...
) -> NamedTuple{(:data,), <:Tuple{NamedTuple{(:ϵ, :σ), <:Tuple{Any, Any}}}}

Fields:

ψ: Material Model

test or tests: A single test or vector of tests. This is used to predict the response of the model in comparison to the experimental data provided.

ps: Model parameters to be used.

Index