Base

Types

BammannChiesaJohnsonPlasticity.AbstractBCJMetalModel — Type

Sub-type for BCJ-models specific to metals.

BammannChiesaJohnsonPlasticity.AbstractBCJMetalTest — Type

Sub-type for BCJ-models specific to metals.

BammannChiesaJohnsonPlasticity.BCJMetalDataEntry — Type

Store vectors of strain, ϵ, and stress, σ data.

BammannChiesaJohnsonPlasticity.BCJMetalStrainControl — Type

Stucture for strain-controlled loadings of metals for temperature, θ; strain rate, ϵ̇; final strain, ϵₙ; number of loading increments, N; loading direction, loaddir ∈ {:tension, :compression, :torsion}.

BammannChiesaJohnsonPlasticity.Bammann1990Modeling — Type

Structure for viscoplasticity model with loading conditions and material properties. Here, uses the effective strain rate based on applied strain rate and loading direction.

BammannChiesaJohnsonPlasticity.Bammann1990Modeling — Method

Bammann1990Modeling(conditions, μ)

Use loading conditions and material properties to construct viscoplasticity model which assumes a Poisson's Ratio of 0.5. Here, μ is the shear modulus.

Functions

BammannChiesaJohnsonPlasticity.update — Method

Use the equations from [1]. Note: though not explicitly listed in paper, temperature equations h = C₁₅ * exp(C₁₆ * θ) and H = C₁₇ * exp(C₁₈ * θ) are included.

Index

BammannChiesaJohnsonPlasticity.AbstractBCJMetalModel

BammannChiesaJohnsonPlasticity.AbstractBCJMetalTest

BammannChiesaJohnsonPlasticity.BCJMetalDataEntry

BammannChiesaJohnsonPlasticity.BCJMetalStrainControl

BammannChiesaJohnsonPlasticity.Bammann1990Modeling

BammannChiesaJohnsonPlasticity.Bammann1990Modeling

BammannChiesaJohnsonPlasticity.update