OptimizationBCJPlasticityExt

This extension is meant to provide additional functionality to the base package with the intent of performing optimization studies with the Optimization.jl and LossFunctions.jl packages to calibrate material model constants against experimental data. This interface with Optimization.jl is largely accomplished by overloading three functions from ContinuumMechanicsBase.jl with the newly defined types in the BCJPlasticity.jl package:

parameters,

parameter_bounds, and

MaterialOptimizationProblem

Most significant of these overloads is that for parameter_bounds wherein the default value for lb of the returned tuple, which represents the lower bounds of possible values during optimization, is constrained to be $[0, \infty)$ to maintain physical admissibility and self-consistency in the constitutive and ISV rate equations.

Types

Functions

ContinuumMechanicsBase.parameter_bounds — Method

Set lower bounds to zero for physical admissibility.

Index